fix(supabase): fix all Kotlin compilation errors in Supabase integration
The Supabase integration code was written against an assumed API that
didn't match the actual Supabase Kotlin SDK 3.5.0. The following
compilation errors were reported by GitHub Actions (CodeQL run #11):

1. SupabaseClientProvider.kt — HttpLoggingInterceptor unresolved
   The okhttp3.logging.HttpLoggingInterceptor class requires the
   com.squareup.okhttp3:logging-interceptor artifact which isn't in
   the project. Removed the debug logging interceptor entirely —
   it's non-essential and the project already has Timber for logging.

2. SupabaseModule.kt — auth/postgrest/storage/realtime/functions
   unresolved
   The plugin accessors (client.auth, client.postgrest, etc.) are
   extension properties that need explicit imports:
     import io.github.jan.supabase.auth.auth
     import io.github.jan.supabase.postgrest.postgrest
     import io.github.jan.supabase.storage.storage
     import io.github.jan.supabase.realtime.realtime
     import io.github.jan.supabase.functions.functions

3. ParamsBuilder.kt — putJsonNull unresolved
   putJsonNull is only available on JsonObjectBuilder, not in custom
   classes. Replaced with JsonNull directly:
     map[key] = JsonNull

4. CloudPlaylistRepository.kt — eq("deleted_at", null) type mismatch
   and Returning.REPRESENTATION not found
   - eq() doesn't accept null — used isNull("deleted_at") instead.
   - delete() doesn't take a 'returning' parameter — used select()
     inside the request lambda to return deleted rows.

5. HistorySyncRepository.kt + LibrarySyncRepository.kt — range()
   parameter type mismatch
   range() expects Long parameters, not Int. Cast with .toLong().

6. StorageRepository.kt — contentType expects ContentType? not String
   Used io.ktor.http.ContentType.parse(mimeType) to convert.

7. SyncManager.kt — PreferenceKeys unresolved, 'get' unresolved,
   'dao' unresolved
   - PreferenceKeys is a file with top-level vals, not an object.
     Changed import to the specific val: LastEpsilonSyncKey.
   - dataStore.get() extension function needed explicit import:
     import com.epsilonmusic.app.utils.get
   - database.dao.method() changed to database.method() because
     MusicDatabase implements DatabaseDao via 'by delegate.dao'
     delegation, so methods are called directly on the database
     instance.

Also fixed in previous commit: removed custom debug signing config
that referenced ~/.android/debug.keystore (doesn't exist on CI runners).

No streaming, playback, metadata, lyrics, Room, Supabase architecture,
Firebase, Discord, or UI changes.

+++ app/src/main/kotlin/com/epsilonmusic/app/supabase/repository/HistorySyncRepository.kt
@@ -63,7 +63,7 @@ class HistorySyncRepository @Inject constructor(
     suspend fun listRecentlyPlayed(limit: Int = 50): List<RecentlyPlayedDto> = try {
         postgrest.from("recently_played").select {
             order("last_played_at", Order.DESCENDING)
-            range(0, (limit - 1).coerceAtLeast(0))
+            range(0L, (limit - 1).coerceAtLeast(0).toLong())
         }.decodeList<RecentlyPlayedDto>()
     } catch (e: Exception) {
         Timber.w(e, "listRecentlyPlayed failed")
@@ -73,7 +73,7 @@ class HistorySyncRepository @Inject constructor(
     suspend fun listListeningHistory(limit: Int = 200): List<com.epsilonmusic.app.supabase.model.ListeningHistoryDto> = try {
         postgrest.from("listening_history").select {
             order("played_at", Order.DESCENDING)
-            range(0, (limit - 1).coerceAtLeast(0))
+            range(0L, (limit - 1).coerceAtLeast(0).toLong())
         }.decodeList<com.epsilonmusic.app.supabase.model.ListeningHistoryDto>()
     } catch (e: Exception) {
         Timber.w(e, "listListeningHistory failed")
